products/api_serializers.py
from rest_framework import serializers
from .models import Product

class ProductSerializer(serializers.ModelSerializer):
    category = serializers.StringRelatedField()

    class Meta:
        model = Product
        fields = ["id", "name", "slug", "category",
                  "short_description", "price", "old_price",
                  "stock", "image"]