requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
django-debug-toolbar==6.1.0
djangorestframework==3.16.1
idna==3.11
pillow==12.0.0
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.4
stripe==14.0.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
